move moderateCommentForNote(Note note, CommentStatus newStatus, Activity activity) from CommentActions to NotificationsUtils
