+++ WordPress/src/main/java/org/wordpress/android/ui/notifications/utils/NotificationsUtils.java
@@ -1,10 +1,12 @@
 package org.wordpress.android.ui.notifications.utils;
 
+import android.app.Activity;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.graphics.Typeface;
 import android.graphics.drawable.Drawable;
 import android.os.Build;
+import android.os.Parcelable;
 import android.preference.PreferenceManager;
 import android.text.Layout;
 import android.text.Spannable;
@@ -18,6 +20,7 @@ import android.view.View;
 import android.widget.TextView;
 
 import com.android.volley.VolleyError;
+import com.cocosw.undobar.UndoBarController;
 import com.google.android.gcm.GCMRegistrar;
 import com.google.gson.Gson;
 import com.google.gson.internal.StringMap;
@@ -31,6 +34,12 @@ import org.wordpress.android.R;
 import org.wordpress.android.WordPress;
 import org.wordpress.android.datasets.ReaderPostTable;
 import org.wordpress.android.models.AccountHelper;
+import org.wordpress.android.models.CommentStatus;
+import org.wordpress.android.models.Note;
+import org.wordpress.android.ui.comments.CommentActions;
+import org.wordpress.android.ui.notifications.NotificationEvents.NoteModerationFailed;
+import org.wordpress.android.ui.notifications.NotificationEvents.NoteModerationStatusChanged;
+import org.wordpress.android.ui.notifications.NotificationEvents.NoteVisibilityChanged;
 import org.wordpress.android.ui.notifications.NotificationsDetailActivity;
 import org.wordpress.android.ui.notifications.blocks.NoteBlock;
 import org.wordpress.android.ui.notifications.blocks.NoteBlockClickableSpan;
@@ -44,6 +53,10 @@ import org.wordpress.android.util.helpers.WPImageGetter;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
+
+import javax.annotation.Nonnull;
+
+import de.greenrobot.event.EventBus;
 
 public class NotificationsUtils {
 
@@ -451,5 +464,65 @@ public class NotificationsUtils {
 
     public static boolean spannableHasCharacterAtIndex(Spannable spannable, char character, int index) {
         return spannable != null && index < spannable.length() && spannable.charAt(index) == character;
+    }
+
+    private static void showUndoBarForNote(final Note note, final CommentStatus status, final Activity activity) {
+        new UndoBarController.UndoBar(activity)
+                .message(status == CommentStatus.TRASH ? R.string.comment_trashed : R.string.comment_spammed)
+                .listener(new UndoBarController.AdvancedUndoListener() {
+                    @Override
+                    public void onHide(Parcelable parcelable) {
+                        // Deleted notifications in Simperium never come back, so we won't
+                        // make the request until the undo bar fades away
+                        CommentActions.moderateCommentForNote(note, status,
+                                new CommentActions.CommentActionListener() {
+                                    @Override
+                                    public void onActionResult(boolean succeeded) {
+                                        if (!succeeded) {
+                                            EventBus.getDefault().post(new NoteVisibilityChanged(note.getId(), false));
+                                            EventBus.getDefault().post(new NoteModerationFailed(note.getId(), status));
+                                        }
+                                    }
+                                });
+                    }
+
+                    @Override
+                    public void onClear(@Nonnull Parcelable[] token) {
+                        //noop
+                    }
+
+                    @Override
+                    public void onUndo(Parcelable parcelable) {
+                        EventBus.getDefault().post(new NoteVisibilityChanged(note.getId(), false));
+                    }
+                }).show();
+    }
+
+    /**
+     * Moderate a comment from a WPCOM notification.
+     * Broadcast EventBus events on update/success/failure and show an undo bar if new status is Trash or Spam
+     */
+    public static void moderateCommentForNote(final Note note, final CommentStatus newStatus, final Activity activity) {
+        if (newStatus == CommentStatus.APPROVED || newStatus == CommentStatus.UNAPPROVED) {
+            note.setLocalStatus(CommentStatus.toRESTString(newStatus));
+            note.save();
+            EventBus.getDefault().post(new NoteModerationStatusChanged(note.getId(), true));
+            CommentActions.moderateCommentForNote(note, newStatus,
+                    new CommentActions.CommentActionListener() {
+                        @Override
+                        public void onActionResult(boolean succeeded) {
+                            EventBus.getDefault().post(new NoteModerationStatusChanged(note.getId(), false));
+                            if (!succeeded) {
+                                note.setLocalStatus(null);
+                                note.save();
+                                EventBus.getDefault().post(new NoteModerationFailed(note.getId(), newStatus));
+                            }
+                        }
+                    });
+        } else if (newStatus == CommentStatus.TRASH || newStatus == CommentStatus.SPAM) {
+            EventBus.getDefault().post(new NoteVisibilityChanged(note.getId(), true));
+            // Show undo bar for trash or spam actions
+            showUndoBarForNote(note, newStatus, activity);
+        }
     }
 }
